refractor - separate controller class
