+++ back_end/auth/modules/auth/controllers/GrantTokenController.js
@@ -1,44 +1,11 @@
-const clientService = require("./services/ClientService");
-const codeService = require("./services/CodeService")
-const tokenService = require("./services/TokenService")
-const userMarkerService = require("./services/UserMarkerService")
-const userService = require('./services/UserService')
-const scopeService = require('./services/ScopeService')
-const helpers = require('../../helpers');
-const randomstring = require("randomstring");
+const codeService = require("../services/CodeService")
+const tokenService = require("../services/TokenService")
+const userService = require('../services/UserService')
+const scopeService = require('../services/ScopeService')
+const helpers = require('../../../helpers');
 const jwt = require('jsonwebtoken')
-const axios = require('axios');
-const responseTrait = require('../../traits/responseTrait')
+const responseTrait = require('../../../traits/responseTrait')
 const otpGenerator = require('otp-generator')
-
-exports.authCodeGrant = async (req, res) => {
-    try {
-        const data = req.body
-        const markedUser = await userMarkerService.getMarkedUser(data.user_id)
-        // generate code
-        const code = helpers.Generator.generateCode({
-            user_id: data.user_id,
-            client_id: data.client_id,
-            scope: data?.scope,
-            redirect_uri: data.redirect_uri,
-            created_at: Math.floor(Date.now() / 1000),
-        })
-
-        // save code for checking
-        await codeService.saveAuthCode(code, data.client_id, data.user_id, process.env.AUTH_CODE_EXP)
-
-        return responseTrait.ResponseSuccess(res, {
-            code: code,
-            user_id: data.user_id,
-            state: data.state == null ? null : data.state,
-            otp: !markedUser || markedUser.is_checked ? false : true,
-        })
-
-    } catch (error) {
-        console.log(error)
-        return responseTrait.ResponseInternalServer(res)
-    }
-}
 
 exports.tokenGrant = async (req, res) => {
     let id_token = ''
@@ -200,111 +167,6 @@ exports.tokenGrant = async (req, res) => {
                 id_token: id_token,
                 id_token_pub_key: id_token_pub_key,
             },
-        })
-    } catch (error) {
-        console.log(error)
-        return responseTrait.ResponseInternalServer(res)
-    }
-}
-
-exports.logout = async (req, res) => {
-    const data = req.body
-    try {
-        // const authorization = req.get('Authorization')
-        const authorization = data.Authorization
-        let arr = authorization.split(" ")
-        const access_token = arr[1];
-        const decoded = jwt.decode(access_token)
-
-        await tokenService.destroyAccessToken(decoded.jti, decoded.sub)
-        await tokenService.destroyRefreshToken(decoded.jti, decoded.sub)
-
-        return responseTrait.ResponseSuccess(res, null)
-    } catch (error) {
-        console.log(error)
-        return responseTrait.ResponseInternalServer(res)
-    }
-}
-
-exports.logoutAll = async (req, res) => {
-    const data = req.body
-    try {
-        // const authorization = req.get('Authorization')
-        const authorization = data.Authorization
-        let arr = authorization.split(" ")
-        const access_token = arr[1];
-        const decoded = jwt.decode(access_token)
-
-        await tokenService.destroyAllKey(decoded.sub)
-
-        return responseTrait.ResponseSuccess(res, null)
-    } catch (error) {
-        console.log(error)
-        return responseTrait.ResponseInternalServer(res)
-    }
-}
-
-exports.sendOtp = async (req, res) => {
-    const data = req.body
-    const user_id = data.user_id
-
-    if (!user_id || user_id == '') {
-        return responseTrait.ResponseInvalid(res)
-    }
-
-    const marked_user = await userMarkerService.getMarkedUser(user_id)
-    if (!marked_user || marked_user.is_checked) {
-        return responseTrait.ResponseInvalid(res)
-    }
-
-    // call api send otp
-    try {
-        const {data} = await axios.post(`${process.env.IDEN_URL}/api/iden/otp/send`, {
-            user_id: user_id,
-        }, {
-            headers: {
-                'Content-Type': 'application/x-www-form-urlencoded'
-            }
-        })
-        if (data.status_code != 200) {
-            return responseTrait.Response(res, 502, "send otp failed!", null)  
-        }
-        return responseTrait.ResponseSuccess(res, data.data)
-    } catch (error) {
-        console.log(error)
-        return responseTrait.ResponseInternalServer(res)
-    }
-}
-
-exports.authenticateOtp = async (req, res) => {
-    const data = req.body
-    const user_id = data.user_id
-    const otp = data.otp
-    const fingerprint = data?.fingerprint
-
-    if (!user_id || user_id == '' ||
-        !otp || otp == '') {
-        return responseTrait.ResponseInvalid(res)
-    }
-
-    try {
-        const {data} = await axios.post(`${process.env.IDEN_URL}/api/iden/otp/authen`, {
-            otp: otp,
-            user_id: user_id,
-        }, {
-            headers: {
-                'Content-Type': 'application/x-www-form-urlencoded'
-            }
-        })
-        const check  = data.data.check
-        if (check) {
-            const unMarkedUser = await userMarkerService.unMarkedUser(user_id, fingerprint)
-            if (!unMarkedUser) {
-                return responseTrait.ResponseInternalServer(res)
-            }
-        }
-        return responseTrait.ResponseSuccess(res, {
-            check: check
         })
     } catch (error) {
         console.log(error)
