+++ back_end/role/modules/role/controllers/PermissionController.js
@@ -1,15 +1,5 @@
-const permissionService = require("./services/PermissionService")
-const scopeService = require("./services/ScopeService")
-const responseTrait = require('../../traits/responseTrait')
-
-exports.getScopesFromRoleId = async (req, res) => {
-    const roleId = req.params.roleId
-    const scopes = await scopeService.getScopesFromRoleId(roleId)
-
-    return responseTrait.ResponseSuccess(res, {
-        scopes: scopes ? scopes : ''
-    })
-}
+const permissionService = require("../services/PermissionService")
+const responseTrait = require('../../../traits/responseTrait')
 
 exports.checkPermission = async (req, res) => {
     try {
